+++ hallo_eltern_cli/commands/mdacommand.py
@@ -15,17 +15,25 @@ class MdaCommand(EmailCommand):
 
         self._process_all = args.process_all
 
+        self._mda_command = args.mda_command
+
     @classmethod
     def register_subparser(cls, subparsers):
         parser = register_command_class(
-            cls, subparsers, 'feeds messages into procmail')
+            cls, subparsers,
+            'feeds messages into a message delivery agent '
+            '(procmail, maildrop, ...)')
 
         parser.add_argument('--process-all',
                             action='store_true',
                             help='process all (even already seen) messages')
 
+        parser.add_argument('--mda-command',
+                            default='/usr/bin/procmail',
+                            help='The command to start the MDA with')
+
     def process_email(self, email):
-        subprocess.run(["/usr/bin/procmail"],
+        subprocess.run([self._mda_command],
                        input=str(email),
                        text=True,
                        check=True)